Move column tests to separate file
